fix interactive tables not working
